Allow setting suggested latency
Suggested latency parameters are the latencies requested when a stream
is first opened. They may range from zero to infinity.

When not specified, the default low output latency for the selected
device will be used. The suggested latency can be used to override the
default and may be required to lower the latency to the absolute minimum
possible.

Ultimately, PortAudio will limit the actual latency within implementable
bounds, but there is no guarantee these are workable, so testing and
benchmarking are always recommended.